refactor(PurchaseCountInputView) : Change the package of PurchaseCountINputView
